Isolated Exclusive Groups feature's code in respective files.
This bring Exclusive Groups in line with Linked Forms structure, where features are isolated as much as possible to their own header/source.
