+++ SPID/src/LookupConfigs.cpp
@@ -1,5 +1,6 @@
 #include "LookupConfigs.h"
 #include "LinkedDistribution.h"
+#include "ExclusiveGroups.h"
 
 namespace INI
 {
@@ -45,43 +46,6 @@ namespace INI
 			//string::replace_all(newValue, "NOT ", "-");
 
 			return newValue;
-		}
-
-		std::optional<RawExclusiveGroup> parse_exclusive_group(const std::string& a_key, const std::string& a_value, const std::string& a_path)
-		{
-			if (a_key != "ExclusiveGroup") {
-				return std::nullopt;
-			}
-
-			const auto sections = string::split(a_value, "|");
-			const auto size = sections.size();
-
-			if (size < 2) {
-				logger::warn("IGNORED: ExclusiveGroup must have a name and at least one Form Filter: {} = {}"sv, a_key, a_value);
-				return std::nullopt;
-			}
-
-			auto split_IDs = distribution::split_entry(sections[1]);
-
-			if (split_IDs.empty()) {
-				logger::warn("ExclusiveGroup must have at least one Form Filter : {} = {}"sv, a_key, a_value);
-				return std::nullopt;
-			}
-
-			RawExclusiveGroup group{};
-			group.name = sections[0];
-			group.path = a_path;
-
-			for (auto& IDs : split_IDs) {
-				if (IDs.at(0) == '-') {
-					IDs.erase(0, 1);
-					group.rawFormFilters.NOT.push_back(distribution::get_record(IDs));
-				} else {
-					group.rawFormFilters.MATCH.push_back(distribution::get_record(IDs));
-				}
-			}
-
-			return group;
 		}
 
 		std::pair<Data, std::optional<std::string>> parse_ini(const RECORD::TYPE& typeHint, const std::string& a_value, const std::string& a_path)
@@ -317,8 +281,7 @@ namespace INI
 
 				for (auto& [key, entry] : *values) {
 					try {
-						if (const auto group = detail::parse_exclusive_group(key.pItem, entry, truncatedPath); group) {
-							exclusiveGroups.emplace_back(*group);
+						if (ExclusiveGroups::INI::TryParse(key.pItem, entry, truncatedPath)) {
 							continue;
 						}
